+++ packages/create-eclesia-indexer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # create-eclesia-indexer
 
+## 1.0.2
+
+### Patch Changes
+
+- efd78ac: Fix comet 38 events
+- 8631218: Polling fix
+
 ## 1.0.1
 
 ### Patch Changes